+++ CHANGELOG.md
@@ -37,7 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       - Delete columns
       - Split long tables into two
 
-## [1.1.0] - 2024-??-??
+## [1.1.0] - 2024-09-30
 
 ### Added